Rename CombineFn -> combinefn
