+++ learning/katas/go/core_transforms/combine/combinefn/task-remote-info.yaml
