+++ learning/katas/go/core_transforms/combine/lesson-info.yaml
@@ -19,5 +19,5 @@
 
 content:
 - simple_function
-- CombineFn
+- combinefn
 - Combine PerKey
